Overview: Changed Map.Size() back to uint. Details: 1. Changed all implementations to use uintptr to store size instead of uint64. This provides better support for 32 bit machine. This is also reasonable. 2. Slightly changed README.md TODO: 1. Make all the helper functions private.

+++ Maps/Map.go
@@ -26,7 +26,7 @@ type ExtendedMap[K any, V any] interface {
 	Map[K, V]
 	//HasKey is a convenient alias for `_,x:=M.Load(K)`
 	HasKey(K) bool
-	Size() uint64
+	Size() uint
 	//Take an arbitrary key value pair from the Map.
 	Take() (K, V)
 	//TakePtr is the pointer variant of Take.
